// packages/backend/src/mxcad/enums/mxcad-return.enum.d.ts
export declare enum MxUploadReturn {
    kOk = "ok",// 成功
    kErrorParam = "errorparam",// 参数错误
    kChunkAlreadyExist = "chunkAlreadyExist",// 分片已存在
    kChunkNoExist = "chunkNoExist",// 分片不存在
    kFileAlreadyExist = "fileAlreadyExist",// 文件已存在
    kFileNoExist = "fileNoExist",// 文件不存在
    kConvertFileError = "convertFileError",// 文件转换错误
    kQuotaExceeded = "quotaExceeded"
}
//# sourceMappingURL=mxcad-return.enum.d.ts.map